New Plugin: clock
This is an analog clock that can be used to show and change a time.
As it is implemented with SVG this plugin is also an example of how to include SVGs, change them through JavaScript (including bus updates) and make parts of the SVG react to user interaction.

TODO:
- make scalable and use colspan + rowspan
- implement the knx_encode for the DPT
- better design of the SVG
- change UI and SVG to show a little clock ("preview") and on click pop up a bigger one for changing the setting
- change SVG design based on visu design (e.g. dynamically recolor parts)
- visually hide the handles
- add digital time to clock (to distinguish AM/PM, but also for those that prefer that and to show the exact time)