New feature: popup events
This will call a "close" event when a popup is being clicked.
This close event can be also called externally to close a popup.
Or you can listen externally to a close event for your own actions at that time. E.g. a diagram could listen to it to stop data fetching.