Quick fix for metal design.
Opening the (normal) demo config with the metal design in Chrome didn't
show the main screen as it was too wide and the browser placed it to the
bottom.

This patch fixes the issue by forcing a recalculation.

A real fix would be to follow though the page creation, especially when
the navbars are active, and handle the sized. This is kept for future
development...