Enhancement: gacefully handle nonexistent plugins
When the contig contains a plugin that does not exists a message will
be sent to the console (to help debugging for the user) but then
the visu will continue to load.

Note: when the plugin will not only be loaded but also be used the
usual "unknown widget" will appear there.